LayoutTests/http/tests/security/module-correct-mime-types-expected.txt

CONSOLE MESSAGE: line 2: Executed script with MIME type: 'application/ecmascript'.
CONSOLE MESSAGE: line 2: Executed script with MIME type: 'application/javascript'.
CONSOLE MESSAGE: line 2: Executed script with MIME type: 'application/x-ecmascript'.
CONSOLE MESSAGE: line 2: Executed script with MIME type: 'application/x-javascript'.
CONSOLE MESSAGE: line 2: Executed script with MIME type: 'text/ecmascript'.
CONSOLE MESSAGE: line 2: Executed script with MIME type: 'text/javascript'.
CONSOLE MESSAGE: line 2: Executed script with MIME type: 'text/javascript1.0'.
CONSOLE MESSAGE: line 2: Executed script with MIME type: 'text/javascript1.1'.
CONSOLE MESSAGE: line 2: Executed script with MIME type: 'text/javascript1.2'.
CONSOLE MESSAGE: line 2: Executed script with MIME type: 'text/javascript1.3'.
CONSOLE MESSAGE: line 2: Executed script with MIME type: 'text/javascript1.4'.
CONSOLE MESSAGE: line 2: Executed script with MIME type: 'text/javascript1.5'.
CONSOLE MESSAGE: line 2: Executed script with MIME type: 'text/jscript'.
CONSOLE MESSAGE: line 2: Executed script with MIME type: 'text/livescript'.
CONSOLE MESSAGE: line 2: Executed script with MIME type: 'text/x-ecmascript'.
CONSOLE MESSAGE: line 2: Executed script with MIME type: 'text/x-javascript'.
Test module scripts run with correct mime types.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Module is not executed yet.
Module scripts run with correct mime types.
PASS successfullyParsed is true

TEST COMPLETE
